chore: unify session service
